Added total records despite pagination

+++ ElasticErrorRates.Core/Models/ElasticResponse.cs
@@ -4,7 +4,7 @@ namespace ElasticErrorRates.Core.Models
 {
     public class ElasticResponse<T> where T : class
     {
-        public int TotalRecords { get; set; }
+        public long TotalRecords { get; set; }
         public IEnumerable<T> Records;
     }
 }
